VCV-150: Finalize refactor
