+++ src/features/operations/ai-performance/components/specimen-confusion-matrix.tsx
@@ -4,7 +4,7 @@ import {
     formatMatrixPercentage,
     getMatrixCellPresentation,
     integerCountFormatter,
-} from '../utils/specimen-confusion-matrix-helpers';
+} from '../utils/ai-performance-helpers';
 import {
     Table,
     TableBody,
