+++ src/features/operations/ai-performance/utils/ai-performance-helpers.ts
