+++ src/features/operations/geographical-summary/components/create-specimen-marker-icon.ts
@@ -2,7 +2,7 @@ import L from 'leaflet';
 import {
     ANOPHELES_COLOR,
     ANOPHELES_THRESHOLD,
-} from '@/features/operations/geographical-summary/utils/site-marker-data';
+} from '@/features/operations/geographical-summary/utils/geographical-summary-helpers';
 
 const MARKER_RADIUS_MIN = 8;
 const MARKER_RADIUS_MAX = 26;
