+++ src/features/operations/geographical-summary/utils/geographical-summary-helpers.ts
